# Support modify column/add bloom filter/add bitmap index for table with primary key