--- a/be/src/agent/heartbeat_server.h
+++ b/be/src/agent/heartbeat_server.h
@@ -42,7 +42,6 @@ class HeartbeatServer : public HeartbeatServiceIf {
 public:
     explicit HeartbeatServer(TMasterInfo* master_info);
     ~HeartbeatServer() override = default;
-    ;
 
     virtual void init_cluster_id();