--- a/be/src/agent/utils.h
+++ b/be/src/agent/utils.h
@@ -34,7 +34,6 @@ class MasterServerClient {
 public:
     MasterServerClient(const TMasterInfo& master_info, FrontendServiceClientCache* client_cache);
     virtual ~MasterServerClient() = default;
-    ;
 
     // Reprot finished task to the master server
     //
@@ -66,9 +65,7 @@ private:
 class AgentUtils {
 public:
     AgentUtils() = default;
-    ;
     virtual ~AgentUtils() = default;
-    ;
 
     // Use rsync synchronize folder from remote agent to local folder
     //