--- a/be/src/runtime/thread_resource_mgr.h
+++ b/be/src/runtime/thread_resource_mgr.h
@@ -87,7 +87,6 @@ public:
     class ResourcePool {
     public:
         virtual ~ResourcePool() = default;
-        ;
         // Acquire a thread for the pool.  This will always succeed; the
         // pool will go over the quota.
         // Pools should use this API to reserve threads they need in order