--- a/be/src/storage/storage_engine.cpp
+++ b/be/src/storage/storage_engine.cpp
@@ -66,23 +66,6 @@
 #include "util/time.h"
 #include "util/trace.h"
 
-using apache::thrift::ThriftDebugString;
-
-using std::back_inserter;
-using std::copy;
-using std::inserter;
-using std::list;
-using std::map;
-using std::nothrow;
-using std::pair;
-using std::priority_queue;
-using std::set;
-using std::set_difference;
-using std::string;
-using std::stringstream;
-using std::vector;
-using strings::Substitute;
-
 namespace starrocks {
 
 StorageEngine* StorageEngine::_s_instance = nullptr;
@@ -230,7 +213,7 @@ Status StorageEngine::_init_store_map() {
 }
 
 void StorageEngine::_update_storage_medium_type_count() {
-    set<TStorageMedium::type> available_storage_medium_types;
+    std::set<TStorageMedium::type> available_storage_medium_types;
 
     std::lock_guard<std::mutex> l(_store_lock);
     for (auto& it : _store_map) {
@@ -404,8 +387,7 @@ std::vector<DataDir*> StorageEngine::get_stores_for_create_tablet(TStorageMedium
         }
     }
     //  TODO(lingbin): should it be a global util func?
-    std::random_device rd;
-    srand(rd());
+    std::srand(std::random_device()());
     std::shuffle(stores.begin(), stores.end(), std::mt19937(std::random_device()()));
     return stores;
 }
@@ -715,14 +697,14 @@ OLAPStatus StorageEngine::_start_trash_sweep(double* usage) {
         *usage = *usage > curr_usage ? *usage : curr_usage;
 
         OLAPStatus curr_res = OLAP_SUCCESS;
-        string snapshot_path = info.path + SNAPSHOT_PREFIX;
+        std::string snapshot_path = info.path + SNAPSHOT_PREFIX;
         curr_res = _do_sweep(snapshot_path, local_now, snapshot_expire);
         if (curr_res != OLAP_SUCCESS) {
             LOG(WARNING) << "failed to sweep snapshot. path=" << snapshot_path << ", err_code=" << curr_res;
             res = curr_res;
         }
 
-        string trash_path = info.path + TRASH_PREFIX;
+        std::string trash_path = info.path + TRASH_PREFIX;
         curr_res = _do_sweep(trash_path, local_now, curr_usage > guard_space ? 0 : trash_expire);
         if (curr_res != OLAP_SUCCESS) {
             LOG(WARNING) << "failed to sweep trash. [path=%s" << trash_path << ", err_code=" << curr_res;
@@ -797,7 +779,7 @@ void StorageEngine::_clean_unused_txns() {
     }
 }
 
-OLAPStatus StorageEngine::_do_sweep(const string& scan_root, const time_t& local_now, const int32_t expire) {
+OLAPStatus StorageEngine::_do_sweep(const std::string& scan_root, const time_t& local_now, const int32_t expire) {
     OLAPStatus res = OLAP_SUCCESS;
     if (!FileUtils::check_exist(scan_root)) {
         // dir not existed. no need to sweep trash.
@@ -806,9 +788,9 @@ OLAPStatus StorageEngine::_do_sweep(const string& scan_root, const time_t& local
 
     try {
         for (const auto& item : std::filesystem::directory_iterator(scan_root)) {
-            string path_name = item.path().string();
-            string dir_name = item.path().filename().string();
-            string str_time = dir_name.substr(0, dir_name.find('.'));
+            std::string path_name = item.path().string();
+            std::string dir_name = item.path().filename().string();
+            std::string str_time = dir_name.substr(0, dir_name.find('.'));
             tm local_tm_create;
             memset(&local_tm_create, 0, sizeof(tm));
 
@@ -822,7 +804,7 @@ OLAPStatus StorageEngine::_do_sweep(const string& scan_root, const time_t& local
             // try get timeout in dir name, the old snapshot dir does not contain timeout
             // eg: 20190818221123.3.86400, the 86400 is timeout, in second
             size_t pos = dir_name.find('.', str_time.size() + 1);
-            if (pos != string::npos) {
+            if (pos != std::string::npos) {
                 actual_expire = std::stoi(dir_name.substr(pos + 1));
             }
             VLOG(10) << "get actual expire time " << actual_expire << " of dir: " << dir_name;
@@ -871,8 +853,7 @@ void StorageEngine::add_unused_rowset(const RowsetSharedPtr& rowset) {
     auto rowset_id = rowset->rowset_id().to_string();
 
     std::lock_guard lock(_gc_mutex);
-    auto it = _unused_rowsets.find(rowset_id);
-    if (it == _unused_rowsets.end()) {
+    if (_unused_rowsets.find(rowset_id) == _unused_rowsets.end()) {
         rowset->set_need_delete_file();
         rowset->close();
         _unused_rowsets[rowset_id] = rowset;
@@ -931,7 +912,7 @@ OLAPStatus StorageEngine::obtain_shard_path(TStorageMedium::type storage_medium,
     return res;
 }
 
-OLAPStatus StorageEngine::load_header(const string& shard_path, const TCloneReq& request, bool restore,
+OLAPStatus StorageEngine::load_header(const std::string& shard_path, const TCloneReq& request, bool restore,
                                       bool is_primary_key) {
     DataDir* store = nullptr;
     {