… be/src/storage/tablet_meta_manager.cpp
… be/src/storage/tablet_updates.h
… be/src/storage/vectorized/convert_helper.cpp
… be/src/storage/vectorized/delta_writer.cpp
… be/src/storage/vectorized/range.h
… be/src/storage/vectorized/schema_change.cpp
… be/src/storage/vectorized/schema_change.h
… be/src/storage/vectorized/tablet_reader.cpp
… be/src/storage/vectorized/tablet_reader.h
… be/src/storage/wrapper_field.cpp
… be/test/storage/tablet_mgr_test.cpp
… be/test/storage/tablet_updates_test.cpp
… be/test/storage/vectorized/schema_change_test.cpp
… fe/fe-core/src/main/java/com/starrocks/alter/SchemaChangeHandler.java
… fe/fe-core/src/main/java/com/starrocks/alter/SchemaChangeJobV2.java
… fe/fe-core/src/main/java/com/starrocks/analysis/AggregateInfo.java
… fe/fe-core/src/main/java/com/starrocks/analysis/Analyzer.java
… fe/fe-core/src/main/java/com/starrocks/analysis/StmtRewriter.java
… fe/fe-core/src/main/java/com/starrocks/catalog/HashDistributionInfo.java
… fe/fe-core/src/main/java/com/starrocks/catalog/PartitionInfo.java
… fe/fe-core/src/main/java/com/starrocks/metric/Metric.java
… fe/fe-core/src/main/java/com/starrocks/qe/QueryDetail.java
… fe/spark-dpp/src/main/java/com/starrocks/load/loadv2/dpp/SparkDpp.java
… fe/spark-dpp/src/main/java/com/starrocks/load/loadv2/dpp/SparkRDDAggregator.java