Table with primary key only support add/delete column when doing schema change.
The pull request is to support more schema change including modify column type/add bloom filter/add bitmap index/drop bitmap index.
I make a performance test on SSB 100G using Query 1.4. Add a bloomfilter index on `s_address` which the cardinality is 20000. After the bloomfilter, the data expansion ratio is 0.408%. And the latency like the following.
```
without bloomfilter:  14.63 sec(the first time),  0.19 sec(the second time).
with bloomfilter:      3.10 sec(the first time),  0.03 sec(the second time).
Data is read from disk upon first time, read from OS page cache upon second time.
```