@chaoyli 2021-11-16
Does this commit can also support modify column ?