@chaoyli 2021-11-17
> > Does this commit can also support modify column ?
> 
> Yes.

@Linkerist How about change the order of primary key?