@chaoyli 2021-11-17
If add modify column, you should add more unit test to check the feature.